VideoPress: Create videopress rest api v1 token class
